vout: spu: rework channel registration
Every registered channels need to be unregistered. The first 3 OSD channels
don't need to be unregistered since they are automatically registered.

sput_t now use a vector of struct spu_channel to keep its channels. This struct
will contain parameters specific to a channel like the clock/delay/rate.

The channel_id is now a ssize_t, -1 being invalid, 0, 1, 2 being the OSD
channels, and [3; SSIZE_MAX] being registered spu channels.

ES_OUT_VOUT_FLUSH_OVERLAY is renamed to ES_OUT_VOUT_DEL_OVERLAY since this
control is only used to delete an overlay.

Refs #22273

+++ include/vlc_subpicture.h
@@ -165,7 +165,7 @@ struct subpicture_t
 {
     /** \name Channel ID */
     /**@{*/
-    int             i_channel;                    /**< subpicture channel ID */
+    ssize_t         i_channel;                    /**< subpicture channel ID */
     /**@}*/
 
     /** \name Type and flags
